Solve PermutationCheck for O(N)
Implemented a bunch of rules to cover various cases in determining if
an array is a permutation.

For evaluation results, see:
  https://codility.com/demo/results/trainingNUSCKP-MDD/

This code should be refactored to make it more comprehensible.

Should try out some other approach, such as checking if the the total
sum of array matches the sum of all integers in range [1-max(input)],
but that alone can't serve as a reliable method.