Solve FrogRiverOne for O(N)
Picked the earliest times for each leaf, since they stay in place.
After that I just needed to find the largest of the result set, since
that's the time the last location gets covered.

If the number of locations covered is not the same as the number of
jumps needed to cross the river, poor Mr. Toad The Frog can't reach the
other side.

For result evaluation, see:
  https://codility.com/demo/results/trainingFNPRWE-RF2/