Solve DivisionCounter for O(N)
This was a tough one to crack!

For a given range, the max int that's divisible by K is the first lower
integer value of range end divided by K (it's the floor of range end / K).

The min int that's divisible by K is first greater integer value of range
start divided by K (that's the ceil of range start / K).

To count the number of divisible integers within a range, we need to
subtract the range start result (0 .. range start) from range end result
(0 .. range end).

The trick is to add 1 to the final result, because we're subtracting the
range start result, which is supposed to be included in the overall result.

For result evaluation, see:
  https://codility.com/demo/results/trainingMM8XZV-UTD/