Solve MaxCounters for O(N)
Changed the algorithm a bit in order to avoid hitting a timeout.

Full optimization achieved using a temp variable to hold transient
states of maximal counter values, and setting the counters values
when appropriate.

This allows us to fill the result array with max values just once,
instead of per loop turn.