Optimize code to exit early if possible
We're keeping track of covered positions, so we can exit the loop early
(cross the river) - as soon as all the positions are covered. This way
(in a not worst case scenario) we don't need to reach the end of the
input list, potentially reducing execution time for large inputs.